# apps/docs/public/openapi/v1.0.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sendero Agent Tools",
    "version": "1.0.0",
    "description": "The Sendero tool registry — a single surface for travel booking, treasury rebalancing, OCR, trip-document verification, and guest-escrow settlement.\n\n## Authentication\n\nEvery call is authenticated with a Clerk-issued API key (`Authorization: Bearer ak_…`). Mint one at [`/dashboard/settings/api-keys`](https://www.sendero.travel/dashboard/settings/api-keys). Keys are tenant-scoped; the server derives both `tenantId` and the service-account `userId` from the key.\n\n## Dispatch shape\n\nEvery tool is callable via `POST /api/agent/dispatch` with `{ tool, args, tenantId }` in the body. The same tools are also exposed via MCP at `/api/mcp` for agent runtimes that prefer the MCP protocol.\n\n## Pricing\n\nEach call bills a per-tool nanopayment in micro-USDC via x402. See [`packages/tools/src/pricing.ts`](https://github.com/tcxcx/sendero/blob/main/packages/tools/src/pricing.ts) for the catalog. SaaS plan tiers grant a discount on the nanopayment rate.",
    "contact": {
      "name": "Sendero Developer Experience",
      "url": "https://docs.sendero.travel"
    }
  },
  "servers": [
    {
      "url": "https://www.sendero.travel",
      "description": "Production"
    },
    {
      "url": "https://preview.sendero.travel",
      "description": "Preview"
    }
  ],
  "components": {
    "securitySchemes": {
      "ClerkApiKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ak_… (Clerk-issued)",
        "description": "Mint a production key at /dashboard/settings/api-keys. Sandbox keys are auto-issued on tenant creation and don't move real USDC."
      }
    },
    "schemas": {
...
      "ConfirmBookingInput": {
        "type": "object",
        "required": ["bookingId", "costMicroUsdc", "itineraryHash", "vendorAddress"],
        "properties": {
          "bookingId": {
            "type": "string"
          },
          "costMicroUsdc": {
            "type": "string",
            "description": "Supplier net rate in micro-USDC (decimal string)."
          },
          "markupBps": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10000
          },
          "markupMicroUsdc": {
            "type": "string"
          },
          "override": {
            "type": "object",
            "properties": {
              "reason": {
                "type": "string",
                "enum": ["ceiling_acknowledged"]
              },
              "acknowledgedMicroUsdc": {
                "type": "string"
              }
            },
            "required": ["reason", "acknowledgedMicroUsdc"]
          },
          "itineraryHash": {
            "type": "string"
          },
          "itineraryCID": {
            "type": "string"
          },
          "vendorAddress": {
            "type": "string"
          },
          "escrowAddress": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/agent/dispatch#confirm_booking": {
      "post": {
        "operationId": "confirm_booking",
        "summary": "Agent path: pin tenant pricing policy snapshot, compute markup breakdown (supplier cost + agency markup + Sendero take), persist the brea…",
        "description": "Agent path: pin tenant pricing policy snapshot, compute markup breakdown (supplier cost + agency markup + Sendero take), persist the breakdown to the Booking row, and encode the commitBookingV2 userOp. Three-recipient release: vendor + agency + operator. Operator submits the userOp; this tool only encodes. Charges the per-call x402 fee + the Sendero take in one MeterEvent.",
        "tags": ["Settlement"],
        "security": [
          {
            "ClerkApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tenantId", "tool", "args"],
                "properties": {
                  "tenantId": {
                    "type": "string",
                    "description": "The calling tenant's id (tenant_...)."
                  },
                  "userId": {
                    "type": "string",
                    "description": "Optional subject for the call; omitted in service-account mode (we use svc:${keyId} automatically)."
                  },
                  "tool": {
                    "type": "string",
                    "enum": ["confirm_booking"],
                    "description": "Must be exactly `confirm_booking` for this operation."
                  },
                  "args": {
                    "$ref": "#/components/schemas/ConfirmBookingInput"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tool executed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "result": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "latencyMs": {
                      "type": "number"
                    },
                    "priceMicroUsdc": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key."
          },
          "402": {
            "description": "Tenant spend cap exceeded."
          },
          "404": {
            "description": "Tool or tenant not found."
          },
          "500": {
            "description": "Tool execution failed."
          }
        }
      }
    },
...
  }
}